-Changes some variables name in AgendaRepositoryImpl.kt -Removes ByteArrayHelperFunctions.kt and combines it with UriByteConverters.kt -Removes some private functions in EventDetailViewModel.kt to setup values directly
