+++ app/src/main/java/com/andrew/tasky/agenda/data/agenda/AgendaRepositoryImpl.kt
@@ -57,16 +57,17 @@ class AgendaRepositoryImpl(
             }
         }
 
-        emit(
-            reminders.combine(tasks) { reminderFlow, taskFlow ->
-                reminderFlow + taskFlow
-            }.combine(events) { reminderTaskFlow, eventsFlow ->
-                reminderTaskFlow + eventsFlow
+        // Added underscore to get rid of name shadowing error
+        emitAll(
+            reminders.combine(tasks) { _reminders, _tasks ->
+                _reminders + _tasks
+            }.combine(events) { _remindersAndTasks, _events ->
+                _remindersAndTasks + _events
             }.map {
                 it.sortedBy { agendaItem ->
                     agendaItem.startDateAndTime
                 }
-            }.first()
+            }
         )
 
         syncAgendaItems()
@@ -122,17 +123,6 @@ class AgendaRepositoryImpl(
                 Log.e("Update Agenda of Date", "Unknown Error")
             }
         }
-        emitAll(
-            reminders.combine(tasks) { reminderFlow, taskFlow ->
-                reminderFlow + taskFlow
-            }.combine(events) { reminderTaskFlow, eventsFlow ->
-                reminderTaskFlow + eventsFlow
-            }.map {
-                it.sortedBy { agendaItem ->
-                    agendaItem.startDateAndTime
-                }
-            }
-        )
     }
 
     override suspend fun syncAgendaItems() {
